Niz: fix inheriting image attributes
We have two kinds of images, cloud and zuul.  Some image attributes
can be set at the provider level for convenince and inherited.  But
if a provider has both types of images, then we might copy a value
that is only relevant for a cloud image to a zuul image (or vice versa).

To correct this, check which type of image we are dealing with and
only copy the relevant attributes.

Change-Id: I13d7347be3838b8a7fa2c53b62cd7add8af61fd0