Fix race in OIDC cache init
It is possible for the cache playback from init to call the
postCacheHook before our list of cache_updated_listeners is
initialized.  Fix this by initializing it before calling the
superclass init.

It is okay if it is initialized empty; the current call site
is in zuul-web and it will add the listener before it first
loads data from the cache.

Change-Id: Iaea1b5fa0d214721811b976dad88dcd02b58bedd