AWS: only import from same s3 region
If we're performing an import from s3, that can only be accomplished
within the same region.  If we are given an s3 url for an image,
check to see if it's in the same region as the target upload; if not,
fall back to the download/upload method.

Experimentally, boto is able to download objects from a bucket in
a region other than the connection, so the existing code which
tries to download the image from any connection can remain as-is
(there is no need to add the same check to the downloadUrl method).

Change-Id: Ia847ca95788c9c85b4884cfbb47dc12efe708d72