Fix image build event race and add more image api tests
This change does two things:

1) It adds tests that verify that in the case where images are
   managed in a central repo, admins may not use the web api of
   the wrong tenant (ie, a tenant that is not actually managing
   the builds of these images) to trigger or delete a build or
   upload.

In expanding the tests to include more than one tenant, a previously
existing race condition was more reliably observed, so it is fixed:

2) The following race sequence could happen:

[1] launcher triggers builds for 2 missing images
[2] scheduler begins reporting the image buildset (2 jobs)
[2] scheduler adds image build artifact #1 to registry
[1] launcher receives signal from zk watch that the image registry
    was updated and checks for missing images
[1] launcher observes that image #2 is still missing and submits
    an image build trigger
[2] scheduler adds image build artifact #2 to registry
[2] scheduler removes queue item
[2] scheduler processes trigger event, begins new build of
    image #2

To resolve this, we note that the addition of multiple image artifacts
is a critical section.  We adjust the creation order so that the
sequence is now:

* Create artifact #1 with state=None
* Create artifact #2 with state=None
* Create upload #1 and update artifact #1 to ready
* Create upload #2 and update artifact #2 to ready

The critical section is now bounded by the condition of having any
IBAs with state=None.  We now check for that within the launcher and wait
for it to clear before we decide if any images are missing.

Change-Id: Iad1b68cf8c9cf6bd5f13dab0471e7bf5fd290fbf