Catch exception in the executor exit handler
This change prevents the executor from being stuck when the
exit handler raises an exception, for example when
the ZooKeeper connection is lost.
This change is particularly useful for graceful handler when
the supervisor will wait multiple hour before sending a SIGKILL
signal. In that case, we should retry the graceful process.

Change-Id: I9c5c39f1862859fb3486626723314d5ec078fbe2