Handle ZooKeeper connection loss in the graceful stop handler
This change prevents the executor from being stuck when a graceful
stop is initiated when the ZooKeeper connection is lost.
In that case, the pause handler may raise an exception when setting
the component status, and if we don't catch the exception the
executor doesn't proceed with the stop logic.

Moreover, if any exceptions is raised by the graceful process,
this change calls the stop handler to ensure proper terminaison.

Change-Id: I9c5c39f1862859fb3486626723314d5ec078fbe2