Fix missing review dismissal due to same timestamp
In some cases a user or bot might submit a 'CHANGES_REQUESTED' and
'DISMISSED' review so fast that the reviews have the same submit
timestamp. So far the Github driver did not correctly dismiss the
previous review.

Since the Github REST API returns reviews in chronological order [0], we
don't need to check the `grantedOn` (in the response `submitted_at`)
timestamp and can instead just process them in order.

[0] https://docs.github.com/en/rest/pulls/reviews?apiVersion=2022-11-28#list-reviews-for-a-pull-request

Change-Id: Ied906b41d7b8b07e32d7f73c7d0a1a0dbb5072f1