Use formatted timestamp for nodeset state_time in API
In the REST API, all of our timestamps are formatted strings
except for the state_time of nodes.  As we switch from nodepool
to nodepool-in-zuul, let's use the formatted version of the
timestamp for NIZ records for consistency with the rest of the
system.  This avoids an API change for any current users and
will only introduce the new value when they start to use NIZ.

Updated the web UI to handle both versions.

Change-Id: I9052ea8803f5f2d62f3523695c071224da9906d7