Do not update gerrit change cache with old data
An issue was observed with a change series in gerrit when using
cherry-pick.  The sequence of events on two schedulers was:

[2] tell gerrit to merge 1,1
[1] receive a comment-added event for 1,1
[1] perform a gerrit query for 1,1 and receive state=NEW
[2] perform the isMerged query for 1,1 and receive state=MERGED
[2] update 1,1 in the change cache with is_merged=True
[1] update 1,1 in the change cache with is_merged=False

This race requires that scheduler #1 perform a gerrit query and store
the information locally, while scheduler #2 merges the change and
performs a query on the same change and saves it to the change cache,
all before scheduler #1 actually saves its data to the change cache.
This means that scheduler #1 writes old data to the change cache and
flips the is_merged flag back to false.

This can occur easily if there are multiple related changes.  In this
case, scheduler #1 may spend quite some time querying other changes
while scheduler #2 only has to query a single change.

Once queue processing resumes after the isMerged query completes,
scheduler #2 will remove the change at the head, then consider the
next change.  It will see a change that depends on the first change,
and it will see that is_merged on the first change is false, therefore
it will remove the dependency.

So far, we have described a sequence of events that could happen with
any merge strategy, so why does this only appear to be a problem with
cherry-pick?  That's because right after the comment-added event is
sent, gerrit sends the change-merged event.  In most merge-modes, that
will cause us to query the change a third time and then set the
is_merged flag back to true.  So there is only a very small window
where this race condition can have an effect with most merge modes.

But with the cherry-pick mode, the change-merged event is emitted for
a new patchset, and it is the change cache entry for this new patchset
where is_merged is set to true.  The original patchset is left with
is_merged=false.

To correct this, we update the gerrit driver so that when it is about
to update the change cache with data from a query, it compares the
ltime of the query data to the ltime already set on the change object.
If another scheduler has updated the change cache and that has
propagated to our change object, and our data are older than what is
on the change object, we will skip the update.

The change cache itself uses the zookeeper object version system to
ensure that the local change object has the most up-to-date data
before writing it to zk.  The check described above will work together
with this to ensure that, first, if any other scheduler has written
the change to zk, we have that data and the query ltime for it; then
second, that our query ltime is newer than that before we update our
data.

Change-Id: Ic206a93ab79242d2cd9d76c006bed67b1f669b31