Filter out old revesions from needed_by
The gerrit change query may return old revisions of other changes that
depend on a given change.  We currently add them to "needed_by" so that
they might be enqueued behind a change when it is enqueued.  However,
in practice we're never going to enqueue an old revision in the gate
pipeline (at least, it isn't going to merge).  So filter it out early
so that we save a few more gerrit queries.

Change-Id: I34856c4ceab086c672327e5454dc04854290363f