Make unique images provider independent
So far an image upload was only valid for the list of providers it was
created for. Since a provider's endpoint-image-config might be
the same between different providers, we should identify image-uploads by
config hash rather than for which providers it was created.

This way the image will be automatically usable for newly added
providers if the image config is the same.

Also, since an artifact might be used in providers with differing image
configurations we need to derive the image name from the upload UUID
instead of the image artifact UUID. Otherwise we can run into naming
conflicts on the endpoint side.

TODO:
* to be checked if this change is backward compatible with existing
  uploads
* anything to do for cleanup?
* in addition to checking the config hash we probably still need the
  provider list. Otherwise we can't continue using an old image because
  the config changed in some way.

Change-Id: Ia84f6965d0851eea5a78774d2d8018d0c82d3b6a