Make unique image uploads provider independent
So far an image upload was only valid for the list of providers it was
created for. Since a provider's endpoint-image-config might be
the same between different providers, we should identify image-uploads by
config hash as well as for which providers it was created.

This way the image will be automatically usable for newly added
providers if the image config is the same. By also keeping a list
of providers the image was originally uploaded for, we can continue
to use an old image if its config (and hash) changes.

Also, since an artifact might be used in providers with differing image
configurations we need to derive the image name from the upload UUID
instead of the image artifact UUID. Otherwise we can run into naming
conflicts on the endpoint side.

Change-Id: Ia84f6965d0851eea5a78774d2d8018d0c82d3b6a
Co-Authored-By: James E. Blair <jim@acmegating.com>