Fix win_shell with stdin
Change I35c220c1774b1943165075c3730236c78b95b18d introduced an
bug where "win_shell" would produce an error if stdin was supplied.
This was because one of the polymorphic forms of the CreateProcess
method was incorrectly identified as unused.  This method is
responsible for converting the stdin string to bytes.

Restore this method and add the additional arguments required
by Zuul so that we can handle stdin strings.

The win_command module does not have polymorphic CreateProcess
methods and is unaffected.

Change-Id: I3a0662a540090b9bf835ee30279e24cae0aad738