Launcher: defer requests with insufficient quota
We currently accept requests as fast as we can and assign new nodes
to them, and then allow those nodes to wait for quota before we try
to actually start building them.  This means that when at capacity,
the system quickly stabilizes in a state where all requests are
accepted and many nodes are waiting in the start state.  Once a
request is accepted, we no longer consider its priority.  This reduces
the effect of request priority (including relative priority).

To correct this, have the accept processing method defer requests if
the chosen provider is at quota.  But if we have already started
assigning nodes to a request, continue to do so (and if the provider
is at quota, we will continue to wait in the start state).  This lets
multi-node jobs get a foothold in the queue once their first label
is able to be fulfilled.

Change-Id: I494b2d42fff96ab5d26e52296ee03b01e9ed3d5d