Use nodeset alternatives for unreachable hosts
The nodeset alternatives mechanism is designed to allow Zuul to
request cheap, unreliable nodes and, if they are unavailable,
to fall back on more expensive and more reliable nodes.  For example,
spot and on-demand instances.

To handle situations where a nodeset is reclaimed while a job is
running (as opposed to simply a lack of availablity at the time of
request), Zuul will now honor nodeset alternatives when retrying
builds if it retries them due to an unreachable host.

If Zuul has already exhausted the list of alternatives (or there are
none), then Zuul will repeat the previous nodeset request.

Retries that are not caused by unreachable hosts will not use the
next nodeset request alternative and will repeat the previous
request.

Change-Id: I53a868ca0e82bffc609b18a016be2cea868c1ca3