Cleanup leaked nodeset requests
As in nodepool, let's have the scheduler cleanup leaked nodeset
requests.

There is no good choice for where to put this.  We need the complete
list of nodeset request ids that the schedulers see, and also the
list of ids in zookeeper.  The schedulers have easy access to the
former, and the launcher the latter via their cache.

To resolve this, rather than adding the cache to the schedulers,
it will probably result in less zk traffic to simply issue a
get_children call to zookeeper.  This will transfer the whole
list of request ids, but only the ids, and only on one of the
schedulers, and only once an hour (or at startup).

Change-Id: I07c92b51eaa6c90313805b3114983301480b1c3b