Allow for project names to be specified as lists
Zuul configuration already allows users to set a project name to a regex
which applies that project configration to all projects whose name's
match that regex. In some cases relying on a single regex may be
problematic due to the complexity of project names. Add functionality
that allows users to supply names as a list. Either as literal projects
names or a list of regexes or a mix of both.

In either case any project that matches either a literal name or any one
of the regexes listed will have that project configuration applied to
it. This should allow users to use simpler project configurations within
central config projects.

To test this we copy the existing regex project test case and modify the
config to use lists of projects and lists of regexes to cover both
situations.

Change-Id: Ibb8223bbdd470fa324111a4b2bb22cf10b8170e2