Fix windows streaming with FQCN
Live log streaming in windows did not work if the playbook used
the FQCN (ansible.windows.wil_shell or win_command) rather than the
short name (just win_shell or win_command) because Ansible would
look up the action module for the FQCN (which is the "normal"
module) rather than use our custom action modules.

To correct this, we will override the normal module to do two things:
1) update the task vars with the necessary settings to support
   live log streaming
2) explicitly call the win_shell module if ansible.windows.win_shell
   is used (and the same for _command); this will use our custom
   module

The task vars setup code is the same (with one exception) for all
shell and command modules (unix/windows + shell/command) no matter
the name, so that code is extracted into a helper method which is
called from all of the action plugins.  The one execption is that
non-split streams is a backwards-compatible feature only available
for unix command/shell, so that is retained in the original location.

A test to exercise this is added, but can not be run in the gate.
Both the test_win_command and test_win_command_fqcn tests have been
run locally with a windows host.

Change-Id: Ie4fc86e2cfc6ee221add46ef2b6bc2d65573d740