Add NIZ autohold support
This adds support in the executor to mark niz nodes as HOLD, and
also for the scheduler or web components to be able to delete
niz nodes that were held due to autoholds.

Management of autohold requests is entirely in Zuul, but implemented
in the zuul/zk/nodepool.py file.  This is fine for now as long
as we update that to use the launcher client to actually update
the niz nodes.  Later, we can remove the autohold request
methods into the launcher client as well.

Change-Id: If6becfceef1d7eb8e28c83a1987c16dc0f840378