Refresh launcher objects after locking
In case the cache is slightly behind, refresh upload and node objects
after we lock them.  This would also protect against a write outside
of a lock (even though that is what prompted this change, that is a
bug, and we should find and fix those instead of masking them).

Change-Id: I357a017ab35e57d98c60b80f69f8f2dc970b182e