Launcher: handle API errors in request assignment
We generally try to avoid cloud API interactions in the request
assignment loop, but we currently do perform some API calls to
get quota information.

Update those calls so that if we encounter an error, we treat
the provider is full or otherwise unavailable, so that we continue
request processing.

Change-Id: I1b2e050a3a67390dfc46161f019dc5899572eac9