Launcher: remove tenant check in getProviderForNode
getProviderForNode serves two purposes: it should return the
provider for a node if it is assigned to one, and it should
find a matching provider if it is not assigned to one (in the
min-ready case).  The previous code interleaved these checks
such that a node in a tenant that was assigned to a provider
that was later removed from the tenant would cause an exception
to be raised because we could not find that provider in that
tenant.

However, if a node is assigned to a specific provider, we
don't need to match it to the tenant, since the provider is
global in nature.

The method is restructured to check the two cases more explicitly:
assigned first, then min-ready/unassigned fallback.

This will allow providers to cleanup nodes even after they are
removed from certain tenants.

Change-Id: I4c73d7eb1c0c2e388a8bc78dd6ee17990971e65c