Remove ignore_label from getProviderForNode
We can simplify provider assignment for min-ready nodes by
storing the provider on the node when we create the state machine.

Essentially, the min-ready nodes will not be assigned to a provider
until a provider starts creating a server for them, at which point
we will consider that node assigned to that provider.

Change-Id: Ifff1f68d72f3defbac9875745188b9da35779ca2