Assign min-ready nodes to providers
Instead of having min-ready nodes exist unattached to providers,
let's mirror the way we select providers for requests and do the
same to select a provider to which we will assign our min-ready
nodes.  This simplifies things because now nodes will always have
associated providers, and we can also be a little smarter about
which provider will actually end up creating instances for
min-ready nodes.

Change-Id: If83d45ca310757b21c98702a8e76027626576963