Prevent Zuul to send docs to EL reporter if not connected
The connection to the ElasticSearch endpoint is protected and prevent
docs to be sent to the endpoint if Zuul has not successfuly connected
to the endpoint at initialization. Furthermore we handle any other
exceptions when adding docs to the endpoint to avoid buildset error in
case of issue accessing the ElasticSearch endpoint.

Change-Id: If909ca573166db9627492ffa86bad5ce6c823a2b