Prevent Zuul to send docs to EL reporter if not connected
The connection to the ElasticSearch endpoint is protected and prevent docs to
be sent to the endpoint if Zuul is not connected to the endpoint.

Also in case of Zuul was connected to the endpoint but the TLS certficate
expires after the connection was established (what we experimented in
production), or any other exception, we log the exception w/o raising
it to the scheduler to avoid a buildset failure.

Change-Id: If909ca573166db9627492ffa86bad5ce6c823a2b