Launcher: improve upload logs
This removes one unecessarily chatty log line when we're checking
for old images (we repeat the line for every image in every provider,
so it ends up being a lot of lines and nothing usually happens).

It adds some log lines around the image upload job so we have better
visibility into the overall progress of the job and, when we're
waiting for a particular upload to finish, which one we're waiting
for.

The previous commit added a "Finished upload" log line that is
closer to when the upload is actually finished; this commit remove
an identical line which we emitted after waiting for the upload.
Since we're adding a log line at the start of the wait, and there
is a line after all the waits are complete, we no longer need the
old "Finished" line.

Change-Id: Iaee2683b4c2ef4354e6e6b37384a2547fe1f33c5