Launcher: don't update upload timestamp
We currently set the image upload timestamp when we create it
(which happens as soon as we learn about the existence of the build
artifact) and also when we finish the actual upload.  But the actual
upload could happen out of order, perhaps even long after later
artifacts are uploaded.  To avoid issues where Zuul uses an older
artifact because its upload is newer, stop setting the timestamp
when the upload completes.

We still set the state time once we set the upload to ready, so we
will still have an indication of when the upload completes.

Change-Id: I44674192632993a109c9f98a684443aae10bb8f2