Expose launcher locks thorugh the web api
The launcher uses four tree caches: requests, nodes, artifacts,
uploads.  The web server also has those caches so that it can answer
queries about their contents.  It would be useful to be able to
show what component has locked a given artifact, upload, or request,
or node.  To do so, this change does the following:

* Updates the ZuulTreeCache so that the parsePath method can indicate
  whether the cache should fetch the data in the object.  This is so
  that we can fetch the contents of lock contender nodes, but still
  avoid any other unecessary fetches.
* Stores information about the component in the ZKContext that we create
  and use for locking ZKObjects.
* Sets the lock identifier (which is stored in the lock contender node)
  to the component's hostname (retrieved from the ZKContext).
* Updates the LockableZKObjectCache to set the name of the current
  holder (the winning contender) as a non-serialized attribute on
  the object.
* Exposes whether an object is locked and the current lock holder
  through the REST API.

The scheduler was relying on the zk_client duck-typing as a ZKContext
when it acquired some locks; but now that we need more information
from the ZKContext, this has been updated to create an extra context
when acquiring the lock.

Change-Id: I00328433fb6ecf5d965e8a4d5bf492a80b4ee1b4