Replace rather than restart failed uploads
In OpenDev, we observed a failed upload due to a cloud error.  The
launcher correctly reset that upload to PENDING and another launcher
re-attempted the upload later.  It finished without error, but the
resulting image was corrup.  We suspect that multi-part uploads to
the same target image name may not behave correctly.  Rather than
attempt to ensure that it is always safe to restart an upload, let's
just delete the original upload on failure and replace it with a
new record with a new uuid.  This ensures we start the replacement
upload with a clean slate, and can separately delete whatever
traces of the original upload remain.

Change-Id: I89e97f75866cc9be84e7640a43d062ac17d85c1b