Let selectProviders dictate use of ready nodes
Consider a ready node in provider B, and a request for two nodes.
If selectProviders decides that the request should be handled by
provider A, then we will end up nodes from different providers
because:

* The request handler will see the ready node from provider B and
  add it to the request.
* The request handler will then create a new node in provider A
  for the second node.

To avoid this, we might be able to inspect update selectProviders
to favor a provider that has a ready node needed by the request,
but only if that provider can handle the rest of the request.
That method is already quite complex; let's see if a simpler
solution is acceptable.

This change simply ignores ready nodes that aren't in whatever
provider selectProviders had already chosen for that label.  This
may leave some ready nodes unused for longer.  It may also, in
extreme edge cases, avoid providers with ready nodes because
those ready nodes use more quota -- at least until other providers
catch up and selectProviders chooses the provider with the ready
node.

But those seem unlikely to be significant problems in practice,
so let's try to resolve this simply.

Additionally, add the provider information back to ready nodes
that we use so that it appears in status listings and in
the inventory correctly.

Change-Id: Iab1696e02e447c02a994a7ae3230a69affc35caa