Store quota on initial node creation
The request acceptance loop looks at the quota situation to decide
if nodes should be assigned to providers and whether a request should
be accepted.  If many requests arrive at once, it may make decisions
with outdated information.  One contributing factor to that is that
we don't store the quota information on the node until we actually
allocate an instance.  We can improve the accuracy by storing our
initial guess of the quota used based on the label.  If that changes
as the creation process progresses, it will be revised, but having
any data at all should let the request processing loop perform better.

Change-Id: Ib549c2842f5b46563d46ac41431e4d48f5a77018