Disable custom win_ modules when console disabled
Users may use old versions of Windows without the necessary .Net
framework needed to run Zuul's win_shell and win_command modules.  To
avoid breaking these users, setting the zuul_console_disabled variable
will now also disable execution of these modules and switch back to
using the built-in Ansible modules.  This technique could also be
applied to the Posix versions, but this change does not do so, in
order to minimize any behavioral differences.

To accomplish this, we build on earlier work that maps the FQCN module
names to shortened ones.  It is updated to map all of the variants of
win_command/shell to run zuul_win_command/shell modules.  These are
our custom modules, renamed.  If the user sets zuul_console_disabled,
this mapping is turned off and we fall through to the normal case
and use the included Ansible module.

Change-Id: I5b4467e2dff389280e725b3d1f80a4742d1b6072