Skip ZooKeeper components update when ZooKeeper is offline
This change prevents the executor from being deadlock when
initiating a graceful stop while ZooKeeper is not available.

In the manageLoad, the unregister_work is set to false, but
if the component raises a ConnectionLoss exception, then,
another attempt to gracefully stop the executor will not
unregister because the value was already set.

This change ensures that local properties are only registered
after being successfully submitted to ZooKeeper.

Change-Id: Icd52094c25e475182b6fa12e1d378b0b05ff7a97