Fix provider locality on failure
When one node in a multinode request fails, we attempt to preserve
the locality -- launching the next attempt in the "main" provider
for a request.  However, our determination of the main provider
based on pre-existing nodes in the request compared strings to
objects and failed.  Correct that, updating some variable names
to make things more clear in the process.

Change-Id: I9f131842ec560af91c4b693b04ac7e6dbfe1ac1c