Keep orphaned image uploads
An orphaned upload (one where we don't have a matching provider) means
that either

    a) the provider was removed without following the provider removal
       procedure by clearing out the list of images before deleting it

or

    b) that we could not load the provider, e.g. due to a config error.

In both cases we want to keep the upload in case the provider shows up
again, e.g. when the config error is fixed.

So far the upload could not be deleted as long as the provider wasn't
available, leading to an exception. However, a pending delete job could
succeed as soon as the provider was added again.

Change-Id: I5aadeda7f3ebd37b22b050fecf3d1bcde517e073