Always add image-build-name to vars when set
So far we've only added the image build name as a Zuul variable when
image formats were requested. However, this is only the case when the
image is already attached to a provider.

With this change we will always set the Zuul variable if the job
image-build-name attribute is set.

Change-Id: I9fc672ccc05eb4d4f07cebafa59adf3acedc3790