Further fixes to multi-tenant provider locality
We may have multiple python objects for identical providers that
are in different tenants.  When the selectProviders method obtained
a provider object by canonical name, it may have ended up with
an object from a different tenant, which would then not match the
identity test we use to decide if it is present in the list of
available providers.

To correct this, obtain the request's main provider object by
searching for the provider in the list of providers for the current
tenant.

Change-Id: I4b8070812f86e73ae80ac9bc644e4545120cec35