Fix pre-timeout calculation
This change addresses three issues with job pre-timeouts:

We incorrectly calculated the timeout for the main run playbook
if a pre-timeout was set.  This is because we called
"getAnsibleTimeout" twice in that case, and used its output as
input the second time.  Instead, we should just call it once
per playbook, and pass in either the pre or run timeout.

It also adds an explicit error for the case where a user sets
pre-timeout to a value larger than timeout.  Previously that
was reported as a MaxTimeoutError, but the error text for that
doesn't apply to this situation.

Finally, while we do have error checking for a user setting
pre-timeout > timeout on the same job, we don't do any checking
of that if the situation arises due to inheritance.  In that
case, let's silently use the smaller timeout value as the
pre-timeout value as well.  That is the most sensible interpretation
of these conflicting values, and should avoid unecessary run-time
errors.

Change-Id: Ife67f5c097583f4cb9a4dc4bcf984ed83964c523