Fix double assignment of ready nodes
We did not double check that a ready node is still available after
locking it when assigning it to a request.  This could allow multiple
zuul-launchers to race assigning the node to their requests.

Correct that by double checking the status after locking the node.

Also, move the refresh call into the try/finally that unlocks the
node.

Change-Id: Id45ddbe5624c80d2800e2f15e5358786babbabdf