Don't count paused provider time against launch-timeout
If a request is assigned to a provider, but we have "paused" that
provider, it is likely the case that the provider is at quota
and there are no other providers that can handle the request.

In that case, we should not count the time that the provider is
paused against the launch timeout of a node.  To correct this,
set the create_state_machine start_time to be the moment that
we first advance the state machine past the START state.

Change-Id: Ie23179131247d3c0bf4ef1ad085c20f1a55da42e