Consider requested nodes when accepting requests
We want to avoid accepting a nodeset request and assigning nodes
to a provider with insufficient quota.  When making that decision,
we currently do not consider nodes in the "requested" state as
contributing to that provider's quota use.  This can lead to
assigning more nodes to a provider than it can handle, which can
make for inefficient distribution since another provider may be
able to handle that node earlier.

To improve this, consider nodes in the "requested" state as
using the provider's quota when making the decision about whether
to send more nodes to that provider.

Change-Id: Ib8e083dda9323df0fce4d4ea0f610939a8b9c062