Add use-nodepool tenant option
This is part of nodepool-in-zuul.

This adds a tenant configuration option, "use-nodepool" which defaults
to true.  True means that the tenant is part of a system that is connected
to nodepool and nodepool will provide some or all of the nodes in the
usual way (this includes acting as a fallback for zuul-launcher).  False
means that all labels used by that tenant will be served by zuul-launcher.

When set to false, if a node request is made for an unknown label, then
Zuul itself will produce a NODE_ERROR result for the build instead of
attempting a nodepool fallback.

Additionally, any references to undefined labels are reported as
configuration errors.  This will help users discover job configuration
errors sooner.  It may increase the number of configuration errors
currently reported in tenants, since previously, any label was presumed
valid.  However, like other invalid reference config errors, these do not
necessarily impede work except in the repo where the error occurs and
can be fixed.  Moreover, they can be suppressed by simply defining
a matching label name (nothing requires the label be attached to a
provider).  Labels that are defined but not attached to providers will,
when use-nodepool is false, result in NODE_ERROR in the same way as
a missing nodepool label.

This is intentionally undocumented at the moment because the NIZ work
is not yet complete.

Change-Id: I33f75e5efc21ded0efbf4db783aac3c8c496a2db