Add compatibility with Gerrit pull-replication
When using Gerrit pull-replication we will get
fetch-ref-replication-scheduled and fetch-ref-replicated events. Handle
those similar to the existing replication events by counting matching
pairs.

In cases where we get both kind of replication events (not sure if this
is possible), we would wait for all matching pairs (or until the
replication timout is exceeded).

Change-Id: Icff694b744892728359129d5b018d5207177fe9f