WIP: Stub provider documentation
This adds a script that can be used to print a rough representation
of the provider schema.  It also adds two stub documentation pages
(derived from this scripts output) for the openstack and aws
provider schemas.

They are not reachable from the main documentation (they must be
directly accessed).  They also do not contain any narrative text,
only the names of the options themselves.

The purpose of this change is to provide a current view of all
of the schema settings as we work to ensure the structure is correct
before we begin documenting them.

Change-Id: Ifd00d6515b292db83c14f123088308ea7f4c313d