Stub provider documentation
This adds a script that can be used to print a representation
of the provider schema in either YAML or RST format.

A "doc" attribute is added to the voluptuous schema keys, so that
we can embed the documentation in the schema definitions in the
source code, and then automatically generate the docs from that
for the various drivers.  Since there will be a lot of duplication
among the drivers, this should help keep things manageable.

A small number of documentation strings are added to the schema
keys in order to exercise the feature and ensure that the keys in
complex structures show up where expected.

We don't necessarily have to only maintain this documentation
automatically; we could simply use this script to generate the
initial skeleton and then embellish the docs in RST format.  But
at least the initial generation will be simplified and consistent.

This change also adds two stub documentation pages (derived from
this script's output) for the openstack and aws provider schemas.

They are not reachable from the main documentation (they must be
directly accessed).  They also do not contain any narrative text,
only the names of the options themselves.

The purpose of this change is to provide a current view of all
of the schema settings as we work to ensure the structure is correct
before we begin documenting them.

Change-Id: Ifd00d6515b292db83c14f123088308ea7f4c313d