Add QuotaCache class
This class will be used to store quota information about endpoints.

Not usage -- that we can get from the node cache.

This will store the upper limits of the endpoint, as well as the
quota usage for specific instance types, volume types, etc.  All of
this is in the service of getting every launcher enough information
to decide how to assign requests, even if a particular launcher can
not execute API calls against certain endpoints.

Change-Id: I40f88fe87ac14eee03b7b63a7ce6e50e7732f114