Update drivers to use QuotaCache
This change, for the most part, updates the aws and openstack
drivers to use the new QuotaCache.  This is done by having the
launcher instruct the providers to write updated data to the
cache on startup.  The getQuota* methods are updated to use
date from the cache instead of directly accessing the cloud.

The information will be updated once each time a launcher
handling that provider starts and once each time a tenant
conntaining that provider is updated.  There is no effective
periodic update with this change.

The quota limits are still subject to the endpoint cache, which
means they may need an extra update cycle to refresh.

The usage information is also subject to the endpoint cache which
may never be updated during the lifetime of the process (therefore,
if the number of cores used by a flavor changes, a restart of at
least one launcher process may be needed to trigger the update).

We may want to change that behavior in the future, but since
the usage information is expected never to change and the limit
information changes very infrequently, it's not important to do
that now.

Change-Id: I9dc92dbc7263c665cc7be56f35f98e62d1b38278