Delete nodeset request lock path
The nodeset request is slightly different than other LockableZKObjects
in that we delete it without holding the lock.  That is, its lifecycle
is:

* created by scheduler
* locked by launcher
* released by launcher
* deleted by scheduler

Another possibility is:

* created by scheduler
* locked by launcher
* deleted by scheduler (cancel request)
* released by launcher

The LockableZKObject class was written assuming that the object would
be deleted while the lock was being held.  For example, ProviderNode
objects behave this way.  As a result, the lock path is only cleaned
up with the lock is released and then only if the object itself was
deleted.

To correct this, we will check whether we hold the lock while deleting,
and if we don't, then we will also delete the lock path.

Lock path deleting is idempotent, so in the case that the scheduler
deletes the request because it's canceled, it will delete the lock path
(which means it will remove the lock from under the launcher).  The
launcher will later attempt to delete the lock, but it no-op becaues
its already deleted.  It will not log an error in this case because
the lock delete method catches NoNodeError.

Change-Id: Ifab3cef4d1706b9ab65727793834ddc0aebdc3c9